Replace girder folder id with dandiset id in URL
- Split FileBrowser.vue into corresponding publish and girder components
- Change file browser URL
- Add store getter for version
